delivery-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
pika==1.3.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
geopy==2.4.1
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
asyncio-mqtt==0.13.0